Update approximations.py
I tried to address two issues. First, I tried to compute the determinant in the log scale for each term. Second, I used `solve()` instead of `inv()` to return the inverse of a matrix. Further, I noticed that there are four terms to get the trace of product of two matrix, so I made some changes.